pyproject.toml
[tool.poetry]
name = "repl_python3_beta_Dhruv-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
python-dotenv = "^0.14.0"
flask = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
